requirements.txt
aiohttp==3.7.4.post0
ansiwrap==0.8.4
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
black==21.6b0
bleach==3.3.0
cachetools==4.2.2
-e git+git@github.com:cal-itp/calitp-py.git@435bc3ebdfb0bc69f6240286b959ba24a85096a9#egg=calitp
certifi==2021.5.30
cffi==1.14.5
cfgv==3.3.0
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
descartes==1.1.0
distlib==0.3.2
docutils==0.17.1
entrypoints==0.3
filelock==3.0.12
fsspec==2021.6.0
future==0.18.2
gcsfs==0.8.0
google-api-core==1.30.0
google-auth==1.31.0
google-auth-oauthlib==0.4.4
google-cloud-bigquery==2.16.1
google-cloud-bigquery-storage==2.4.0
google-cloud-core==1.7.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.38.0
gtfs-realtime-bindings==0.0.7
identify==2.2.10
idna==2.10
importlib-metadata==4.6.0
iniconfig==1.1.1
ipdb==0.13.4
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keyring==23.0.1
kiwisolver==1.3.1
libcst==0.3.19
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
mizani==0.7.3
multidict==5.1.0
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nodeenv==1.6.0
notebook==6.4.0
numpy==1.20.3
oauthlib==3.1.1
packaging==20.9
palettable==3.3.0
pandas==1.1.4
pandas-gbq==0.14.1
pandocfilters==1.4.3
papermill==2.3.3
parso==0.8.2
pathspec==0.8.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pkginfo==1.7.0
plotnine==0.8.0
pluggy==0.13.1
pre-commit==2.13.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
proto-plus==1.18.1
protobuf==3.17.3
protobuf-to-dict==0.1.0
ptyprocess==0.7.0
py==1.10.0
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybigquery==0.7.0
pycparser==2.20
pydata-google-auth==1.2.0
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.4
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
readme-renderer==29.0
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.7.2
scipy==1.7.0
Send2Trash==1.7.1
siuba @ git+https://github.com/machow/siuba.git@ca708d023cffae4208b239f549e6ef30159b61b7
six==1.16.0
SQLAlchemy==1.3.24
statsmodels==0.12.2
tenacity==7.0.0
terminado==0.10.1
testpath==0.5.0
textwrap3==0.9.2
toml==0.10.2
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
twine==3.4.1
typing-extensions==3.10.0.0
typing-inspect==0.7.1
ujson==4.0.2
urllib3==1.26.5
virtualenv==20.4.7
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
yarl==1.6.3
zipp==3.4.1
